fix multiple issues with cancel
Cancel entities were being deleted at the end of every refresh which
caused cancellation to be reversed up the following refresh.

Event cancels now have a foreign key to the calendar rule ID. This is
needed so that, when a rule is disabled, its associated cancellation may
be deleted.

Timestamp values are now stored as seconds since epoch instead of
milliseconds since epoch. This is more consistent with what is
recommended in SQLite documentation.